move armada app into cluster stack
